# docs/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
  /workout-sessions:
    get:
      tags:
      - workout-session-controller
      operationId: listWorkoutSessions
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutSessionDto"
    post:
      tags:
      - workout-session-controller
      operationId: createWorkoutSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSessionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
  /workout-sessions/{id}/exercises:
    post:
      tags:
      - workout-session-controller
      operationId: addWorkoutSessionExercise
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSessionExerciseRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
  /workout-sessions/{id}/attendants:
    post:
      tags:
      - workout-session-controller
      operationId: addWorkoutSessionAttendant
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSessionAttendantRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
  /users:
    get:
      tags:
      - user-controller
      operationId: listUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
    post:
      tags:
      - user-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /workout-sessions/{id}:
    get:
      tags:
      - workout-session-controller
      operationId: getWorkoutSession
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
    patch:
      tags:
      - workout-session-controller
      operationId: updateCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutSessionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkoutSessionDto"
  /users/{id}:
    get:
      tags:
      - user-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
    patch:
      tags:
      - user-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
...
  /membership-types/{id}:
    get:
      tags:
      - membership-type-controller
      operationId: getMembershipType
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipTypeDto"
    patch:
      tags:
      - membership-type-controller
      operationId: updateCard_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipTypeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MembershipTypeDto"
...
  /cards/{id}:
    get:
      tags:
      - card-controller
      operationId: getCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
    patch:
      tags:
      - card-controller
      operationId: updateCard_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
  /users/{id}/chats:
    get:
      tags:
      - user-controller
      operationId: listChats
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatDto"
...
  /workout-sessions/{id}/exercises/{exerciseId}:
    delete:
      tags:
      - workout-session-controller
      operationId: deleteWorkoutSessionExercise
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: exerciseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /workout-sessions/{id}/attendants/{userId}:
    delete:
      tags:
      - workout-session-controller
      operationId: deleteWorkoutSessionAttendant
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
...
components:
  schemas:
    WorkoutSessionRequest:
      type: object
      properties:
        coachUuid:
          type: string
          format: uuid
        hallUuid:
          type: string
          format: uuid
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
      required:
      - coachUuid
      - date
      - description
      - hallUuid
      - title
    Card:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        country:
          $ref: "#/components/schemas/Country"
        cardNumber:
          type: string
        cvv:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        nameOnCard:
          type: string
        postalCode:
          type: string
      required:
      - cardNumber
      - cvv
      - dateOfBirth
      - nameOnCard
      - postalCode
    Country:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        countryName:
          type: string
      required:
      - countryName
    Exercise:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        targetMuscles:
          type: array
          items:
            $ref: "#/components/schemas/TargetMuscle"
          uniqueItems: true
    Hall:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        hallType:
          $ref: "#/components/schemas/HallType"
        hallName:
          type: string
        hallDescription:
          type: string
    HallType:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
    Membership:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        purchaseDate:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        membershipType:
          $ref: "#/components/schemas/MembershipType"
    MembershipType:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        type:
          type: string
        price:
          type: number
          format: float
    TargetMuscle:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
    User:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/UserRole"
        card:
          $ref: "#/components/schemas/Card"
        membership:
          $ref: "#/components/schemas/Membership"
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        email:
          type: string
        phoneNumber:
          type: string
        imageUrl:
          type: string
      required:
      - dateOfBirth
      - firstName
      - lastName
    UserRole:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        role:
          type: string
    WorkoutSessionDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        coach:
          $ref: "#/components/schemas/User"
        hall:
          $ref: "#/components/schemas/Hall"
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: string
        attendants:
          type: array
          items:
            $ref: "#/components/schemas/User"
          uniqueItems: true
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSessionExerciseDto"
    WorkoutSessionExerciseDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        exercise:
          $ref: "#/components/schemas/Exercise"
        reps:
          type: integer
          format: int32
        exerciseOrder:
          type: integer
          format: int32
    WorkoutSessionExerciseRequest:
      type: object
      properties:
        exerciseUuid:
          type: string
          format: uuid
        reps:
          type: integer
          format: int32
        exerciseOrder:
          type: integer
          format: int32
      required:
      - exerciseOrder
      - exerciseUuid
      - reps
    WorkoutSessionAttendantRequest:
      type: object
      properties:
        userUuid:
          type: string
          format: uuid
      required:
      - userUuid
    UserRequest:
      type: object
      properties:
        roleUuid:
          type: string
          format: uuid
        cardUuid:
          type: string
          format: uuid
        membershipUuid:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        email:
          type: string
        phoneNumber:
          type: string
        imageUrl:
          type: string
      required:
      - cardUuid
      - dateOfBirth
      - email
      - firstName
      - imageUrl
      - lastName
      - membershipUuid
      - phoneNumber
      - roleUuid
    UserDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/UserRole"
        card:
          $ref: "#/components/schemas/Card"
        membership:
          $ref: "#/components/schemas/Membership"
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        email:
          type: string
        phoneNumber:
          type: string
        imageUrl:
          type: string
...
    MembershipTypeRequest:
      type: object
      properties:
        type:
          type: string
        price:
          type: number
          format: float
      required:
      - price
      - type
    MembershipTypeDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        type:
          type: string
        price:
          type: number
          format: float
...
    ChatDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ChatParticipantDto"
          uniqueItems: true
    ChatParticipantDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        user:
          $ref: "#/components/schemas/User"
        lastReadDateTime:
          type: string
          format: date-time
...
    CardRequest:
      type: object
      properties:
        countryUuid:
          type: string
          format: uuid
        cardNumber:
          type: string
        cvv:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        nameOnCard:
          type: string
        postalCode:
          type: string
      required:
      - cardNumber
      - countryUuid
      - cvv
      - dateOfBirth
      - nameOnCard
      - postalCode
    CardDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        country:
          $ref: "#/components/schemas/Country"
        cardNumber:
          type: string
        cvv:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        nameOnCard:
          type: string
        postalCode:
          type: string
...